Overhauled - Build against WorldEdit 6.0 and WorldGuard 6 snapshots - Removed junk - Refactored - Added extension helpers - Added event documentation to README
