+++ src/main/java/com/zeoldcraft/chworldguard/abstraction/events/WGRegionChangeEvent.java
@@ -1,4 +1,4 @@
-package com.zeoldcraft.dev.abstraction.events;
+package com.zeoldcraft.chworldguard.abstraction.events;
 
 import java.util.List;
