--- src/main/java/com/zeoldcraft/dev/DataType.java
@@ -1,10 +1,0 @@
-package com.zeoldcraft.dev;
-
-public enum DataType {
-	BOOLEAN,
-	DOUBLE,
-	INT,
-	STRING,
-	PLUGIN,
-	SELF
-}
